+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 - PROJECT : display title of membersUserSet Field 
 - STOCK : filter product according to stochManage = true
 - Stock Move: update future quantity from real quantity in planned stock moves.
+- Invoice : add the possiblity to active/desactive the pdf generation on ventilation for customers or suppliers in invoice application configuration
 
 
 ## Bug Fixes